@zimulala 2020-06-18
/run-common-test